added tests to make sure tableView data source and controller are playing nicely... they are.

+++ iOS-Documents-objc/iOS-Documents-objc/Model Controllers/JLCDocumentController.h
@@ -13,7 +13,7 @@
 
 @property (nonatomic, readonly) NSArray *documents;
 
-@property (nonatomic, class, readonly) int totalDocumentsCreated;
+@property (nonatomic, class, readonly) NSUInteger totalDocumentsCreated;
 
 - (void)createDocumentWithTitle:(NSString *)title
                       text:(NSString *)text;
